gml,xtf,xtf24: fixed export of 0.000001
git-svn-id: file:///C/usr/ce/projects/sf/ili2c/gitmigration/ili2c.svn/ili2c/trunk@1254 22a1f05a-e80c-4436-89a7-166f12dbcbdd

+++ src/ch/interlis/ili2c/generator/Gml32Generator.java
@@ -1104,8 +1104,8 @@ public final class Gml32Generator
 					ipw.println ("<xsd:restriction base=\"xsd:decimal\">");
 				}
 		          ipw.indent ();
-		          ipw.println ("<xsd:minInclusive value=\""+((NumericType)type).getMinimum().doubleValue()+"\"/>");
-		          ipw.println ("<xsd:maxInclusive value=\""+((NumericType)type).getMaximum().doubleValue()+"\"/>");
+		          ipw.println ("<xsd:minInclusive value=\""+((NumericType)type).getMinimum()+"\"/>");
+		          ipw.println ("<xsd:maxInclusive value=\""+((NumericType)type).getMaximum()+"\"/>");
 		          ipw.unindent ();
 	            ipw.println ("</xsd:restriction>");
 			}else{
