+++ src/ch/interlis/ili2c/generator/XSD24Generator.java
@@ -844,8 +844,8 @@ public final class XSD24Generator
 				if(min.getExponent()!=0){
 					ipw.println ("<xsd:restriction base=\"xsd:double\">");
 			          ipw.indent ();
-			          ipw.println ("<xsd:minInclusive value=\""+((NumericType)type).getMinimum().doubleValue()+"\"/>");
-			          ipw.println ("<xsd:maxInclusive value=\""+((NumericType)type).getMaximum().doubleValue()+"\"/>");
+			          ipw.println ("<xsd:minInclusive value=\""+((NumericType)type).getMinimum()+"\"/>");
+			          ipw.println ("<xsd:maxInclusive value=\""+((NumericType)type).getMaximum()+"\"/>");
 			          ipw.unindent ();
 				}else{
 					ipw.println ("<xsd:restriction base=\"xsd:decimal\">");
